requirements_dev.txt
wheel==0.33.6
tox==3.14.0
coverage==4.5.4
ruff==0.3.5
pytest==6.2.4
mypy
pandas-stubs